extract php7 only test suite
