+++ tests/ResourceTest.php
@@ -25,8 +25,8 @@ class ResourceTest extends \PHPUnit_Framework_TestCase
     protected function setUp()
     {
         parent::setUp();
-        $this->resource = (
-        new Injector(new FakeSchemeModule(new ResourceModule('FakeVendor\Sandbox')), $_ENV['TMP_DIR']))->getInstance(ResourceInterface::class);
+        $injector = new Injector(new FakeSchemeModule(new ResourceModule('FakeVendor\Sandbox'), $_ENV['TMP_DIR']));
+        $this->resource = $injector->getInstance(ResourceInterface::class);
     }
 
     public function testManualConstruction()
@@ -153,53 +153,5 @@ class ResourceTest extends \PHPUnit_Framework_TestCase
 }
 ';
         $this->assertSame($expected, (string) $user);
-    }
-
-    public function testShortSyntax()
-    {
-        $ro = $this->resource->get->uri('page://self/index')(['id' => 'koriym']);
-        /* @var $ro ResourceObject */
-        $this->assertInstanceOf(Index::class, $ro);
-        $this->assertSame('koriym', $ro->body);
-    }
-
-    public function testShortSyntaxWithQuery()
-    {
-        $ro = $this->resource->get->uri('page://self/index?id=koriym')();
-        /* @var $ro ResourceObject */
-        $this->assertInstanceOf(Index::class, $ro);
-        $this->assertSame('koriym', $ro->body);
-    }
-
-    public function testShortSyntaxInvoke()
-    {
-        $ro = $this->resource->get->uri('page://self/index?id=koriym')->__invoke(['id' => 'koriym']);
-        $this->assertInstanceOf(Index::class, $ro);
-        $this->assertSame('koriym', $ro->body);
-    }
-
-    public function testShortSyntaxFunction()
-    {
-        $index = $this->resource->get->uri('page://self/index?id=koriym');
-        $ro = $index((['id' => 'koriym']));
-        $this->assertInstanceOf(AbstractRequest::class, $index);
-        $this->assertInstanceOf(Index::class, $ro);
-
-        return $index;
-    }
-
-    /**
-     * @depends testShortSyntaxFunction
-     */
-    public function testShortSyntaxReuseRequest(AbstractRequest $index)
-    {
-        $ro = $index(['id' => 'bear']);
-        $this->assertSame('bear', $ro->body);
-    }
-
-    public function testShortSyntaxFunctionWithDefaultGetMethod()
-    {
-        $ro = $this->resource->uri('page://self/index')();
-        $this->assertInstanceOf(Index::class, $ro);
     }
 }
